+++ slides/Dotazy.tsx
@@ -1,6 +1,6 @@
 import Slide from "@/components/Slide";
 
-const DotazySlide = () => (
+const Dotazy = () => (
   <Slide background='./images/outro.jpg' centered>
     <h1 className='text-white bg-black/70 text-center text-4xl font-extrabold uppercase py-10 backdrop-blur-sm w-full'>
       Dotazy
@@ -8,4 +8,4 @@ const DotazySlide = () => (
   </Slide>
 );
 
-export default DotazySlide;
+export default Dotazy;